validate requested hash matches actual
Do not process requests for hashes that mismatch from what we can serve.
This happens when two servers are running different versions of the code
but a load balancer may not know which one to route the request to. We
need to make sure that the we don't accidentally serve a bad file for a
given hash otherwise CDN may cache incorrect things.

The load balancer should be configured for sticky sessions for website
load balancing to minimize impact of hitting a backend with an unknown
file.